Fixing merge conflict (module refactor related)
